AI-BASE/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
playwright>=1.40.0
openai>=1.3.7
google-api-python-client>=2.108.0
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
requests>=2.31.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1 